https://github.com/uN1ck/FateExtended/projects/1#card-56350746 : Доделываю механзм работы с FatePoint-ами, возможность броска из листа напрямую.
